Commit 15
Issue #1 Pause button

The "pause button" issue was fixed by adding an if statement. This
checks if the game is already paused: if it is paused, it plays and vice
versa.

Issue #2 Mute button

The "mute button" issue was fixed by adding an if statement. This check
if the game is already muted, if so it unmutes and vice versa.

Issue #6 Exit

The "exit issue" was fixed by changing the command from
Application.Exit(); to Application.Quit();